requirements.txt
opentelemetry-proto
grpcio